Specifiy OS X path, moving /usr/local/bin to the front; fix for Homebrew.
Aspirationally it would be nice to have a funciton that will take path
items and "move them to the front"; that is, drop them from the path if
they exist, and then prepend them to the path. This is a bigger problem,
so for now this fixes homebrew so that it overrides the system binaries.