Add an alias for "git status"-like display of a specified commit.
Use a cute alias-shell command trick to create an alias that will show
"git status"-like output for a specified commit. Relies on the caret to
specify first parent, so may be ... interesting with multi-parent
commits. Interestingly, is the exact alias in the stack overflow post
which showed the technique.
  http://stackoverflow.com/questions/3321492/git-alias-with-positional-parameters

Also shorten the alias for the same output between arbitrary commites.